Fixed: link on win32 (MinGw & VS) Fixed: restored some plugins Fixed: warnings--
git-svn-id: file:///.data/svn_convert/svn@8730 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-playground/plugins/_core/commands.cpp
@@ -145,7 +145,7 @@ CMenu *Commands::processMenu(unsigned id, void *param, int key)
                 if (cmd->text.isEmpty())
                     continue;
                 cmdKey = QKeySequence::mnemonic(i18n(cmd->text));
-                if((cmdKey & ~Qt::UNICODE_ACCEL) == key)
+                if((cmdKey & ~Qt::UNICODE_ACCEL) == (unsigned int)key)
                 {
                     cmd->param = param;
                     EventCommandExec eCmd(cmd);
